PR-HELPERS-3SPLIT — split _helpers.py 3-way into domain-named siblings
PR-CLEANUP-6 (v0.99.48) flagged core/agent/loop/_helpers.py as the
lone deferred holdout from the file-name hygiene sweep: the new
Naming CANNOT row forbids _helpers filenames once a caller appears,
but the catch-all here genuinely hosted three unrelated sub-systems
under one umbrella. Naming it to any single domain would have buried
the other two; folding into agent_loop.py would have grown that file
by 170 LOC.

This PR reverses the PR-CLEANUP-1 fold along the actual ownership
line — three new sibling modules, each name-matched to its
sub-system:

- core/agent/loop/_tool_factory.py (95 LOC, new): get_agentic_tools
  + AGENTIC_TOOLS + token-guard constants.
- core/agent/loop/_sub_agent_announce.py (55 LOC, new):
  check_announced_results — OpenClaw Spawn+Announce queue poller.
- core/agent/loop/_planner_dispatch.py (79 LOC, new): try_decompose —
  async planner LLM dispatcher that installs the Plan on
  SessionMetrics.active_plan.

_helpers.py deleted (170 LOC). AgenticLoop._try_decompose +
_check_announced_results delegator methods rewired to the new module
names with full host history preserved in their docstrings
(Tier-3-split sibling → PR-CLEANUP-1 fold → this 3-split). Package
__init__ re-exports AGENTIC_TOOLS / MAX_TOOL_RESULT_TOKENS /
get_agentic_tools from the new _tool_factory module so external
imports only need import-path rewrites (no symbol-name change).

Codex MCP verified (read-only): 7 PASS / 2 FAIL (LOC numbers off,
new files untracked at first commit). Both addressed in this commit.

Resolves the last deferred-item from the v0.99.48 cleanup arc.
