+++ core/agent/loop/__init__.py
@@ -34,7 +34,7 @@ from __future__ import annotations
 # split-into-package refactor preserves this surface.
 from core.config import _resolve_provider as _resolve_provider
 
-from ._helpers import AGENTIC_TOOLS, MAX_TOOL_RESULT_TOKENS, get_agentic_tools
+from ._tool_factory import AGENTIC_TOOLS, MAX_TOOL_RESULT_TOKENS, get_agentic_tools
 from .agent_loop import AgenticLoop
 from .models import AgenticResult, _context_exhausted_message, _ContextExhaustedError
